Lab5_assignment/requirements.txt
openai 
dotenv
pyyaml
aiohttp